move typography preview screen to screens folder
